+++ app/src/main/java/com/example/pyco/data/repositories/CategoriesRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package com.example.pyco.data
+package com.example.pyco.data.repositories
 
 import com.example.pyco.data.daos.CategoryDao
 import com.example.pyco.data.entities.Category
